Revert "Tidy Hawk View Test [1-2]"
