+++ conftest.py
@@ -1,8 +1,6 @@
 import os
 import django
 from django.conf import settings
-from django.core.cache import CacheHandler
-import pytest
 
 # We manually designate which settings we will be using in an environment variable
 # This is similar to what occurs in the `manage.py`
@@ -13,19 +11,3 @@ os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'data.settings_test')
 def pytest_configure():
     settings.DEBUG = False
     django.setup()
-
-
-@pytest.fixture()
-def local_memory_cache(monkeypatch):
-    """Configure settings.CACHES to use LocMemCache."""
-    monkeypatch.setitem(
-        settings.CACHES,
-        'default',
-        {'BACKEND': 'django.core.cache.backends.locmem.LocMemCache'},
-    )
-    cache_handler = CacheHandler()
-    monkeypatch.setattr('django.core.cache.caches', cache_handler)
-
-    yield
-
-    cache_handler['default'].clear()
